--- client/packages/invoices/src/Prescriptions/api/hooks/document/index.ts
@@ -1,5 +1,0 @@
-import { usePrescription } from './usePrescription';
-
-export const Document = {
-  usePrescription,
-};
