+++ client/packages/invoices/src/Prescriptions/api/hooks/index.ts
@@ -1,16 +1,12 @@
-import { Line } from './line';
+// import { Line } from './line';
 
-export const usePrescription = {
-  line: {
-    stockLines: Line.usePrescriptionLine,
-    // rows: Line.usePrescriptionRows,
-    delete: Line.usePrescriptionDeleteLines,
-    deleteSelected: Line.usePrescriptionDeleteSelectedLines,
-    deleteAll: Line.usePrescriptionDeleteAllLines,
-    save: Line.usePrescriptionSaveLines,
-  },
-};
+// export const usePrescription = {
+//   line: {
+//     // stockLines: Line.usePrescriptionLine,
+//   },
+// };
 
 export * from './usePrescriptionList';
 export * from './usePrescriptionSingle';
+export * from './usePrescriptionLines';
 export * from './utils';
