--- client/packages/invoices/src/Prescriptions/api/hooks/line/index.ts
@@ -1,15 +1,0 @@
-import {
-  usePrescriptionDeleteLines,
-  usePrescriptionDeleteSelectedLines,
-  usePrescriptionDeleteAllLines,
-} from './usePrescriptionDeleteLines';
-import { usePrescriptionLine } from './usePrescriptionLine';
-import { usePrescriptionSaveLines } from './usePrescriptionSaveLines';
-
-export const Line = {
-  usePrescriptionLine,
-  usePrescriptionDeleteLines,
-  usePrescriptionDeleteSelectedLines,
-  usePrescriptionDeleteAllLines,
-  usePrescriptionSaveLines,
-};
