+++ client/packages/invoices/src/Prescriptions/api/hooks/utils/index.ts
@@ -1,9 +1,7 @@
-import { usePrescriptionIsDisabled } from './usePrescriptionIsDisabled';
 import { usePrescriptionApi } from './usePrescriptionApi';
 import { usePrescriptionNumber } from './usePrescriptionNumber';
 
 export const Utils = {
-  usePrescriptionIsDisabled,
   usePrescriptionApi,
   usePrescriptionNumber,
 };
