--- client/packages/invoices/src/Prescriptions/api/utils/usePrescriptionIsDisabled.ts
@@ -1,8 +1,0 @@
-import { isPrescriptionDisabled } from '../../../utils';
-import { usePrescription } from '../hooks/document/usePrescription';
-
-export const usePrescriptionIsDisabled = (): boolean => {
-  const { data } = usePrescription();
-  if (!data) return true;
-  return isPrescriptionDisabled(data);
-};
